# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QuizRequestDifficulty = typeof QuizRequestDifficulty[keyof typeof QuizRequestDifficulty];


export const QuizRequestDifficulty = {
  easy: 'easy',
  hard: 'hard',
} as const;

export interface QuizRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  topic: string;
  difficulty?: QuizRequestDifficulty;
}

export interface QuizQuestion {
  question: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  answers: string[];
  topic: string;
}

export interface ExplainRequest {
  question: string;
  selectedAnswer: string;
  topic: string;
}

export interface ExplainResponse {
  explanation: string;
}

export interface TopicsResponse {
  topics: string[];
}

export interface ApiError {
  error: string;
}