# lib/api-zod/src/generated/types/explainRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ExplainRequest {
  question: string;
  selectedAnswer: string;
  topic: string;
}