# lib/api-zod/src/generated/types/quizQuestion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuizQuestion {
  question: string;
  /**
     * @minItems 4
     * @maxItems 4
     */
  answers: string[];
  topic: string;
}