# lib/api-zod/src/generated/types/quizRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizRequestDifficulty } from './quizRequestDifficulty';

export interface QuizRequest {
  /**
     * @minLength 1
     * @maxLength 200
     */
  topic: string;
  difficulty?: QuizRequestDifficulty;
}